Fixed issue #13: "jobs diagram don't show arrows / line"
In the html files, hops are encoded as HTML links (a tags) with a href attribute that identifies the name of the "to" end of the hop. Steps and entries are encoded as divs, and their names are used as the id element.

In JavaScript, the value of the href was taken and used as argument to the document.getElementById() DOM method to retrieve the receiving end of the hop. Howver, this does not work in case the name of the entry/hop contains "special" characters: in those cases, the href attribute value may be URI escaped and contain hex-escaped characters,  whereas the id attribute value remains unescaped.

The fix is to apply decodeURIComponent to the href attribute value before using the value as argument for the document.getElementById() method.