Completely overhauled version. This revision fixes issue #10, "Recurse subdirectories with a job (not the get files step)"
The main points of this revision are:

- it does not rely on the "recurse subdirectories" option for the Get Files input step, which became available only in Kettle 4.0. Instead, this version uses the "Get Subfolder Names" step, which is available at least in Kettle 3.2. (and possibly, earlier versions - i'd have to check)

- the current revision does incremental document generation:
  * .html files in the output tree are deleted if there is no corresponding .kjb/.ktr file in the input tree
  * A .html file is created in the output tree for each .kjb/.ktr file in the input tree with no corresponding .html file in the output tree
  * A .html file is rebuilt in the output tree in case the corresponding .ktr/.kjb file in the input tree has a modified date that exceeds the modified date of the respective .html file

Incremental document generation is arguably a nice to have, not a necessity and one can question the justification for building this - IMO non-trivial - feature. However, since the XSLT transformation for the individual .ktr/.kjb files takes most of the time, this feature will generally result in drastically faster generation of documentation, something which is quite noticeable when documenting hundreds of jobs/transformations (or more)

Note that although subdirectories are created as needed (just like before), empty directories in the output tree are not automatically removed.

- The output directory will be created in case it does not yet exist
- The all-items.html and main.html files are always automatically recreated 
- the index.xml file now contains entries for individual folders. This allows creation of a hierarchical toc. This is one of the features that will be implemented in the very short term.

Note - the checked in version is not yet tested in Linux and will likely break because there is currently nothing in place to make the transform script executable. I will fix this asap (and very shortly). So, please hold back a little longer before checking out this new revision.